// src/app/models/multiple_bills/Bill.model.ts
export interface Bill {
    Payment_method: string;
    Price: number;
    Itbis: number;
    Product: string;
    Product_type: number;
    Quantity: number;
    Subtotal: number;
    Total: number;
    Voucher_type: string;
    Card: string;
    Amount: number;
    Product_id: number;
    RNC: string;
    ClientName: string;
    NCF:string;
}